+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   into the negative range
 * return an error when encoding bad utf to json
 * add conveniences to h.io for stdin/out
+* add support for sendto and recvfrom
 
 ## 0.3.3 - 2016-05-21